Separate AntlrLexer and AntlrParser logic
