+++ Sources/PT.PM.Common/ILanguageHandler.cs
@@ -1,0 +1,7 @@
+namespace PT.PM.Common
+{
+    public interface ILanguageHandler : ILoggable
+    {
+        Language Language { get; }
+    }
+}
